refactor: add custom inputs for mantine useform
