+++ src/components/custom-password-input/index.ts
@@ -1,0 +1,1 @@
+export * from './custom-password-input';
