+++ src/components/custom-text-input/index.ts
@@ -1,0 +1,1 @@
+export * from './custom-text-input';
