--- src/components/form/index.ts
@@ -1,2 +1,0 @@
-export * from './password-input';
-export * from './text-input';
